TEST cl_test divided into 2 tests
So that the output is of reasonable
and travis can display it.
